chore: some changes

+++ src/main/java/org/irmc/industrialrevival/api/items/attributes/Rechargeable.java
@@ -2,6 +2,7 @@ package org.irmc.industrialrevival.api.items.attributes;
 
 import org.bukkit.inventory.ItemStack;
 import org.irmc.industrialrevival.core.utils.EnergyUtil;
+import org.jetbrains.annotations.Range;
 
 public interface Rechargeable extends ItemAttribute {
     double getEnergyCapacity();
@@ -10,7 +11,7 @@ public interface Rechargeable extends ItemAttribute {
         return EnergyUtil.getItemEnergy(item);
     }
 
-    default void setItemEnergy(ItemStack item, long energy) {
+    default void setItemEnergy(ItemStack item, double energy) {
         EnergyUtil.setItemEnergy(item, energy);
     }
